Build stage change functionality.
Changed the "useState" function to a "useReducer" instead to simplify
the code required by the stage change triggering function (aka
handleStageChange) inside the "QuizStageContextProviderComponent".

Function "handleStageChange" is created to execute the dispatch for the
"quizStageState" useReducer. The dispatch currently taked an "id"
parameter coming from the "Start" component and triggers the Reducer
function, which in turn updates the state and changes updates the value
of the "stage" object to be "q1". This value will be used to
conditionally change the display from the "Start" component to the
future "Question" component.

Variable "quizStageCtx" has been added a new key-value pair to have
access to the "handleStageChange" function. The "QuizStageContext"
variable has also been updated acordignly.

Finally, the "Start" component now uses the "QuizStageContext" context
to trigger function "handleStageChange" through the "onClick" event prop
and passes the "start" value to it as the identifier.