Set up displayed content conditionally. Create "Question" component.
Created the "Question" component and gave it some dummy content just to
check if the conditional display works as expected.

Tweaked the "App" component to give it access to the "QuizStageContext".
Now "App" is just another component which uses the "currentStage"
value to conditionally change the component displayed alternating
between "Quiz" and "Results" (for now only accepts "Quiz" but "Results"
will come in the future).

Tweaked the "Quiz" component to give it access to the same context. Quiz
also alternates which sub-component it displays according to the
"currentStage" value (alternating between "Start" and "Question").

Created a new "AppWithProvider" component which is now used to render
the application trough "main.jsx" file. This component contains "App"
wrapped with the context provider.