Build change-between-questions functionality.
Updated Question component by adding a new state "questionNumber",
importing "QuizStageContext" and adding a dummy button below the
previous content.

The new "questionNumber" state will take care of updating the Question
component to change from the current question to the following one. It
has a default value of 1 (for 1st question) and will add up +1 once the
Test Button is clicked by the user and function "handleNextQuestion" is
triggered.

"QuizStageContext" is imported and used to trigger the "stageChange"
function passing the string "questionsStageEnding" as its argument when
the user clicks the same Test Button. This button will trigger function
"handleNextQuestion" by default, except when the "questionNumber" state
reaches a value of 7, when the Button will trigger the "stageChange"
context function instead. This ensures that during the questions stage
only the "Question" component is updated instead of the whole App.

The string values used by the "QuizStageContext" state functions have
been updated to reflect correctly how the function works. Now the state
starts with a default value of "startStage". When the dispatch function
is triggered with the parameter "startStageEnding", the state is updated
to "questionsStage", and the process repeats when the
"questionsStageEnding" value is recieved, updating the state "stage"
value to "resultsStage".
All components that use or produce this values
have been updated accordingly.