Create "solutionTracking state and style selected answer on click.
A second state (solutionTracking) is created within the "Question"
component in order to handle the styling events that inform the user
when it has selected an answer and if was wrong or right.

The "Option" sub-component has been added the "className", "id", and
"onClick" props:
--> "onClick" is used to trigger new function "handleSelectionStyling".
--> "id" is given the answer specific index number within the mapped
array.
-->"className" will have either an empty value or the value of state
variable object "selectionStyling" which will contain the needed class
to change the background color of the selected button to orange.

When the user selects a specific answer, the "handleSelectionStyling"
function triggers, and if the "selectionStyling" object within the
"solutionTracking" state variable has no value, it will update it to
"selected". Also, it will update object "selectedSolution" to the clicked
button's id value, so that this object serves as an identifier for the
current selected answer. Finally, since the "Question" component is
re-executed and the different options mapped once again, only the one
with an id equal to the "selectedSolution" object will be added the
"selected" class, thus displaying an orange background instead.