Build authomatic question handling through progress bar.
ProgressBar component is created to manage the progress html element
which is displayed within the Question component and from now on handles
it instead of the onClick feature used previously.
--> An initial timer of 10000ms is used to handle the time the user has
in order to answer a question, before the "handleNextQuestion" function
is activated in the Question component, thus changing the question.
--> An interval is also used to handle the "remainingTime" state which
controls the value of the progress bar. The state initial value comes
from the Question component (variable "time") and starts with 10000ms to
match the timer.

Question component is updated:
--> "solutionTracking" state has been renamed and changed to
"solutionStyling". Now its initial value is an empty string which will
change to selected when the user has "selected" an aswer and afterwards
either to "correct" or "wrong" after certifying this, and back to an
empty string once the next question is presented, thus restarting the
process.
--> "handleSelectionStyling" is updated. It triggers when the user clicks
an answer an immediately pushes the selected answer to the global arrray
"selectedAnswers". Then sets the TIMER variable to 3000ms (this variable
is the one which is passed onto the the ProgressBar component and sets
its state initial value) and sets "solutionStyling" to "selected", which
styles the selected answer in an orange color. After that, it triggers a
new 3000ms timer (to be paired with the progress bar). When the 3000ms
timer runs out, the TIMER variable changes to 5000ms and "solutionStyling"
value is changed either to "correct" or "wrong", which will style the
answer to green or red accordingly. Finally, another timer of 5000ms is
started to match the current progress bar and after that the TIMER value
updates back to 10000ms (original value) and the "handleNextQuestion"
function is triggered in order to pass to the following one.
--> "handleNextQuestion" is updated. Now it can receive a "skipped"
parameter but will only do so if the user has skipped answering and the
ProgressBar 10000ms timer has gone of completely. If that is the case,
the "skipped" value will be pushed to the "selectedAnswers" array. On
the contrary, if this functions is triggered after the user has selected
an answer, then "solutionStyling" is set back to an empty string. In
both cases the "questionNumber" state is incremented by one, thus passing
to the next question.
--> A "key" prop has been used in the ProgressBar component within
Question in order to update it every time a Question state is updated.
To do so, it has been given "compReset" as a value. compReset is a global
variable with an initial value of zero which will increment with each
change of state but only until "questionNumber" reaches 7.
-->useCallback and useEffect hooks have been used as needed to avoid bugs.