Add answers random shuffling logic.
Created global variable "shuffledAnswers" with a null initial value.
This variable will become an array with the current question answers
randomly shuffled, but only when "solutionStyling" is an empty string,
so that the shuffle only happens with the first component update, after
the change of question (and not when the component updates only for
answer styling purposes). The shuffledAnswers variable is now mapped to
the display.