Display stats/results content within "Summary" component.
For the stats content:

Derived function "createStats" is created to map through the "selected
Answers" array and count the number of correct, wrong and skipped
answers. The function then returns an object containing the "correct
Ratio", "wrongRatio", and "skippedRatio" keys which contain the % of
the corresponding type of answer rounded.

Then const "stats" is created within the component function and executes
the derived "createStats" funcion, thus containing the stats.

Finally, those stats are used directly within the corresponding html
content.

For the results content:

Another derived function "createResults" is used to map through the
"selectedAnswers" array and return an array of objects containing the
"questionIndex", "question", "answer", and "answerClassName" keys with
its corresponding information.

The variable "results" is created within the component function and
executes "createResults" so that it contains the results information.

Finally, the results array is looped by within the ol element thus
creating a list of items containing the needed results information and
giving it the necessary conditional styling.