Disable other answers once the user has selected one.
A new disable prop has been created within the "Option" sub-component in
the "Question" component and passed on to it's button html element. This
disable prop activates the "disabled" status for all the answer buttons
if the "solutionStyling" state value is not "" (if the user has already
selected one of the options).

This avoids buggish styling interactions on selecting more than one
option for the same question.