Improve "handleNextQuestion" function.
"handleNextQuestion" function within "Question" component has been
tweaked to avoid one last unnecessary "questionNumber" state execution.
It has been added a second level of if...else statements within the
first and previous one in order to check if "questionNumber" has reached
7 or not. If that is the case instead of executing the state once more
to change the question, it now executes the "stageChange" context
function which automatically swaps the component displayed from "Quiz"
to "Sumamry".

This changed has allowed to remove the "stageChange" execution from the
parent level within the "Question" function and has also simplified the
conditional checks that are needed to update the "compReset" and
"shuffledAnswers" variables.

It also avoids the appearance of harmless error messages within the
console once the "Quiz" display is replaced by "Summary".